+++ sources/loaders/adl_loader/adl_loader.h
@@ -25,5 +25,5 @@
 //
 */
 
-int ADL_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,unsigned long infotype,void * returnvalue);
+int ADL_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,uint32_t infotype,void * returnvalue);
