+++ sources/loaders/adz_loader/adz_loader.h
@@ -25,4 +25,4 @@
 //
 */
 
-int ADZ_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,unsigned long infotype,void * returnvalue);
+int ADZ_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,uint32_t infotype,void * returnvalue);
