+++ sources/loaders/amigadosfs_loader/amigadosfs_loader.h
@@ -25,6 +25,6 @@
 //
 */
 
-int AMIGADOSFSDK_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,unsigned long infotype,void * returnvalue);
+int AMIGADOSFSDK_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,uint32_t infotype,void * returnvalue);
