+++ sources/loaders/apple2_nib_loader/apple2_nib_loader.h
@@ -25,4 +25,4 @@
 //
 */
 
-int Apple2_nib_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,unsigned long infotype,void * returnvalue);
+int Apple2_nib_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,uint32_t infotype,void * returnvalue);
