+++ sources/loaders/dim_loader/dim_loader.h
@@ -25,4 +25,4 @@
 //
 */
 
-int DIM_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,unsigned long infotype,void * returnvalue);
+int DIM_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,uint32_t infotype,void * returnvalue);
