+++ sources/loaders/fdi_loader/fdi_loader.h
@@ -25,4 +25,4 @@
 //
 */
 
-int FDI_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,unsigned long infotype,void * returnvalue);
+int FDI_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,uint32_t infotype,void * returnvalue);
