+++ sources/loaders/flppcm_loader/flppcm_loader.h
@@ -25,4 +25,4 @@
 //
 */
 
-int FLPPCM_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,unsigned long infotype,void * returnvalue);
+int FLPPCM_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,uint32_t infotype,void * returnvalue);
